+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## staged
 
-- none
+- Fixed bug where `PMD.diag` should have been `LinearAlgebra.diag`
+- Fixed bug in `build_settings_file` where the recursive merge was not overwriting the `settings` variable
 
 ## v2.1.0